[Improve] login page style, fix change teamId bug
