+++ streampark-console/streampark-console-newui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "streampark-webapp",
-  "version": "2.0.0",
+  "version": "1.2.4",
   "author": {
     "name": "streampark",
     "url": "https://streampark.apache.org"
