+++ streampark-console/streampark-console-newui/src/design/color.less
@@ -155,3 +155,5 @@ html {
 
 // Modal
 @modal-mask-bg: fade(@black, 45%);
+
+@background-color-base: #e6f7ff;
