+++ streampark-console/streampark-console-newui/src/design/public.less
@@ -24,8 +24,8 @@
 // =================================
 
 ::-webkit-scrollbar {
-  width: 7px;
-  height: 8px;
+  width: 14px;
+  height: 12px;
 }
 
 // ::-webkit-scrollbar-track {
@@ -33,21 +33,21 @@
 // }
 
 ::-webkit-scrollbar-track {
-  background-color: rgb(0 0 0 / 5%);
+  // background-color: rgb(0 0 0 / 5%);
+  background-color: unset;
 }
 
+::-webkit-scrollbar-thumb:hover,
 ::-webkit-scrollbar-thumb {
-  // background: rgba(0, 0, 0, 0.6);
-  background-color: rgb(144 147 153 / 30%);
-  // background-color: rgba(144, 147, 153, 0.3);
-  border-radius: 2px;
-  box-shadow: inset 0 0 6px rgb(0 0 0 / 20%);
+  background-color: #e6f7ff;
 }
 
-::-webkit-scrollbar-thumb:hover {
-  background-color: @border-color-dark;
+[data-theme="dark"] {
+  ::-webkit-scrollbar-thumb:hover,
+  ::-webkit-scrollbar-thumb {
+    background-color: #2a2a2a;
+  }
 }
-
 // =================================
 // ==============nprogress==========
 // =================================
