+++ streampark-console/streampark-console-newui/src/layouts/default/header/components/user-dropdown/index.vue
@@ -17,8 +17,8 @@
 <template>
   <Dropdown placement="bottomLeft" :overlayClassName="`${prefixCls}-dropdown-overlay`">
     <span :class="[prefixCls, `${prefixCls}--${theme}`]" class="flex">
-      <img :class="`${prefixCls}__header`" :src="getUserInfo.avatar" />
-      <span :class="`${prefixCls}__info hidden md:block`">
+      <!-- <img :class="`${prefixCls}__header`" :src="getUserInfo.avatar" /> -->
+      <span :class="`${prefixCls}__info hidden md:block px-10px`">
         <span :class="`${prefixCls}__name  `" class="truncate">
           {{ getUserInfo.username }}
         </span>
