+++ streampark-console/streampark-console-newui/src/views/base/login/Login.vue
@@ -16,34 +16,12 @@
 -->
 <template>
   <div :class="prefixCls" class="relative w-full h-full px-4">
-    <AppLocalePicker
-      class="absolute text-white top-4 right-4 enter-x xl:text-gray-600"
-      :showText="false"
-      v-if="!sessionTimeout && showLocale"
-    />
-    <AppDarkModeToggle class="absolute top-3 right-7 enter-x" v-if="!sessionTimeout" />
-
-    <span class="-enter-x xl:hidden">
-      <AppLogo :alwaysShowTitle="true" />
-    </span>
-
     <div class="container relative h-full py-2">
       <div class="flex h-full">
-        <div class="hidden min-h-full pl-4 mr-4 xl:flex xl:flex-col xl:w-4/12">
-          <AppLogo class="!w-auto -enter-x" />
-          <div class="my-auto">
-            <div class="mt-10 font-medium text-white -enter-x">
-              <span class="inline-block mt-4 text-3xl"> {{ t('sys.login.signInTitle') }}</span>
-            </div>
-            <div class="mt-5 font-normal text-gray-300 text-md -enter-x">
-              {{ t('sys.login.signInDesc') }}
-            </div>
-          </div>
-        </div>
-        <div class="flex w-full h-full py-5 xl:h-auto xl:py-0 xl:my-0 xl:w-8/12">
+        <div class="flex w-full h-full py-5 xl:h-auto xl:py-0 xl:my-0">
           <div
             :class="`${prefixCls}-form`"
-            class="relative w-full px-5 py-8 mx-auto my-auto rounded-md shadow-md xl:bg-transparent sm:px-8 xl:p-4 xl:shadow-none sm:w-3/4 lg:w-2/4 xl:w-auto enter-x"
+            class="relative w-auto px-8 bg-[rgba(0,0,0,0.5)] py-10 mx-auto my-auto rounded-md shadow-md enter-y"
           >
             <LoginForm />
             <ForgetPasswordForm />
@@ -54,15 +32,9 @@
   </div>
 </template>
 <script lang="ts" setup>
-  // import { computed } from 'vue';
-  import { AppLogo } from '/@/components/Application';
-  import { AppLocalePicker, AppDarkModeToggle } from '/@/components/Application';
   import LoginForm from './LoginForm.vue';
   import ForgetPasswordForm from './ForgetPasswordForm.vue';
-  // import { useGlobSetting } from '/@/hooks/setting';
-  import { useI18n } from '/@/hooks/web/useI18n';
   import { useDesign } from '/@/hooks/web/useDesign';
-  import { useLocaleStore } from '/@/store/modules/locale';
 
   defineProps({
     sessionTimeout: {
@@ -72,9 +44,6 @@
 
   // const globSetting = useGlobSetting();
   const { prefixCls } = useDesign('login');
-  const { t } = useI18n();
-  const localeStore = useLocaleStore();
-  const showLocale = localeStore.getShowPicker;
   // const title = computed(() => globSetting?.title ?? '');
 </script>
 <style lang="less">
@@ -99,10 +68,6 @@
         border: 1px solid #4a5569;
       }
 
-      &-form {
-        background: transparent !important;
-      }
-
       .app-iconify {
         color: #fff;
       }
@@ -117,31 +82,18 @@
   .@{prefix-cls} {
     min-height: 100%;
     overflow: hidden;
-    @media (max-width: @screen-xl) {
-      background-color: #293146;
-
-      .@{prefix-cls}-form {
-        background-color: #fff;
-      }
-    }
-
-    &::before {
+    background: url('/@/assets/images/logo-bg.jpg') no-repeat 50%;
+    background-size: cover;
+    &::after {
       content: '';
-      position: absolute;
-      top: 0;
-      left: 0;
       width: 100%;
-      height: 100%;
-      margin-left: -66%;
-      background: linear-gradient(163.85deg, #1d2129 0%, #00308f 100%);
-      // background-image: url(/@/assets/svg/login-bg.svg);
-      // background-position: 100%;
-      // background-repeat: no-repeat;
-      // background-size: auto 100%;
-
-      @media (max-width: @screen-xl) {
-        display: none;
-      }
+      position: absolute;
+      left: 0;
+      top: 0;
+      bottom: -20px;
+      background: inherit;
+      filter: blur(4px);
+      z-index: 2;
     }
 
     .@{logo-prefix-cls} {
