+++ streampark-console/streampark-console-newui/src/views/flink/app/EditFlink.vue
@@ -225,4 +225,6 @@
     </BasicForm>
   </PageWrapper>
 </template>
-<style scoped></style>
+<style lang="less">
+  @import url('./styles/Add.less');
+</style>
