+++ streampark-console/streampark-console-newui/src/views/flink/app/View.vue
@@ -15,29 +15,23 @@
   limitations under the License.
 -->
 <script lang="ts">
-  import { defineComponent, onMounted, ref, reactive, unref, onUnmounted, watch } from 'vue';
-  import { useI18n } from '/@/hooks/web/useI18n';
+  import { defineComponent, ref, unref, onUnmounted, watch } from 'vue';
   import { useUserStore } from '/@/store/modules/user';
+  import { useAppTableAction } from './hooks/useAppTableAction';
   export default defineComponent({
     name: 'AppView',
   });
 </script>
 <script lang="ts" setup name="AppView">
   import { useTimeoutFn } from '@vueuse/core';
-  import { Row, Col, Tooltip, Badge, Divider, Select, Input, Tag } from 'ant-design-vue';
-  import { fetchAppRecord, fetchDashboard, fetchAppRemove } from '/@/api/flink/app/app';
-  import { fetchFlamegraph } from '/@/api/flink/app/metrics';
-  import { ActionItem, useTable } from '/@/components/Table';
-  import { Icon } from '/@/components/Icon';
-  import { useFlinkApplication } from './hooks/useApp';
-  import { useRouter } from 'vue-router';
-  import { useFlinkAppStore } from '/@/store/modules/flinkApplication';
+  import { Tooltip, Badge, Divider, Tag } from 'ant-design-vue';
+  import { fetchAppRecord } from '/@/api/flink/app/app';
+  import { useTable } from '/@/components/Table';
   import { PageWrapper } from '/@/components/Page';
   import { BasicTable, TableAction } from '/@/components/Table';
   import { AppListRecord } from '/@/api/flink/app/app.type';
-  import { useMessage } from '/@/hooks/web/useMessage';
   import { getAppColumns, launchTitleMap } from './data';
-  import { handleIsStart, handleView } from './utils';
+  import { handleView } from './utils';
   import { useDrawer } from '/@/components/Drawer';
   import { useModal } from '/@/components/Modal';
 
@@ -45,19 +39,10 @@
   import StopApplicationModal from './components/AppView/StopApplicationModal.vue';
   import LogModal from './components/AppView/LogModal.vue';
   import BuildDrawer from './components/AppView/BuildDrawer.vue';
+  import AppDashboard from './components/AppView/AppDashboard.vue';
   import State from './components/State';
 
-  import StatisticCard from './components/AppView/statisticCard.vue';
-
-  const SelectOption = Select.Option;
-  const InputGroup = Input.Group;
-  const InputSearch = Input.Search;
-
-  const { t } = useI18n();
-  const router = useRouter();
   const userStore = useUserStore();
-  const flinkAppStore = useFlinkAppStore();
-  const { createMessage } = useMessage();
 
   const optionApps = {
     starting: new Map(),
@@ -65,52 +50,13 @@
     launch: new Map(),
   };
 
-  const tagsOptions = ref<Recordable>([]);
-  const dashboardLoading = ref(true);
-  const dashBigScreenMap = reactive<Recordable>({});
+  const appDashboardRef = ref<any>();
   const searchText = ref('');
   const tags = ref(undefined);
   const jobType = ref(undefined);
   const userId = ref(undefined);
   const yarn = ref<Nullable<string>>(null);
 
-  // Get Dashboard Metrics Data
-  async function handleDashboard(showLoading: boolean) {
-    try {
-      dashboardLoading.value = showLoading;
-      const res = await fetchDashboard();
-      if (res) {
-        Object.assign(dashBigScreenMap, {
-          availiableTask: {
-            staticstics: { title: 'Available Task Slots', value: res.availableSlot },
-            footer: [
-              { title: 'Task Slots', value: res.totalSlot },
-              { title: 'Task Managers', value: res.totalTM },
-            ],
-          },
-          runningJob: {
-            staticstics: { title: 'Running Jobs', value: res.runningJob },
-            footer: [
-              { title: 'Total Task', value: res.task.total },
-              { title: 'Running Task', value: res.task.running },
-            ],
-          },
-          jobManager: {
-            staticstics: { title: 'JobManager Memory', value: res.jmMemory },
-            footer: [{ title: 'Total JobManager Mem', value: `${res.jmMemory} MB` }],
-          },
-          taskManager: {
-            staticstics: { title: 'TaskManager Memory', value: res.tmMemory },
-            footer: [{ title: 'Total TaskManager Mem', value: `${res.tmMemory} MB` }],
-          },
-        });
-      }
-    } catch (error) {
-      console.error(error);
-    } finally {
-      dashboardLoading.value = false;
-    }
-  }
   const [registerStartModal, { openModal: openStartModal }] = useModal();
   const [registerStopModal, { openModal: openStopModal }] = useModal();
   const [registerLogModal, { openModal: openLogModal }] = useModal();
@@ -173,9 +119,18 @@
     columns: getAppColumns(),
     showIndexColumn: false,
     showTableSetting: true,
+    useSearchForm: true,
     tableSetting: { fullScreen: true, redo: false },
     actionColumn: { dataIndex: 'operation', title: 'Operation', width: 180 },
   });
+  const { getTableActions, getActionDropdown, formConfig } = useAppTableAction(
+    openStartModal,
+    openStopModal,
+    openLogModal,
+    openBuildDrawer,
+    reload,
+    optionApps,
+  );
 
   watch(
     () => userStore.getTeamId,
@@ -194,26 +149,8 @@
     openBuildDrawer(true, { appId: app.id });
   }
 
-  /* Click to edit */
-  function handleEdit(app: AppListRecord) {
-    flinkAppStore.setApplicationId(app.id);
-    if (app.appType === 1) {
-      // jobType( 1 custom code 2: flinkSQL)
-      router.push({ path: '/flink/app/edit_streampark', query: { appId: app.id } });
-    } else if (app.appType === 2) {
-      //Apache Flink
-      router.push({ path: '/flink/app/edit_flink' });
-    }
-  }
-  /* Click for details */
-  function handleDetail(app: AppListRecord) {
-    flinkAppStore.setApplicationId(app.id);
-    router.push({ path: '/flink/app/detail', query: { appId: app.id } });
-  }
-
   /* view */
   async function handleJobView(app: AppListRecord) {
-    console.log('app', app.state, app.optionState);
     // Task is running, restarting, in savePoint
     if ([4, 5].includes(app.state) || app['optionState'] === 4) {
       console.log(app);
@@ -221,172 +158,7 @@
       handleView(app, unref(yarn));
     }
   }
-  /* Click to delete */
-  async function handleDelete(app: AppListRecord) {
-    const hide = createMessage.loading('deleting', 0);
-    try {
-      await fetchAppRemove(app.id);
-      createMessage.success('delete successful');
-      reload();
-    } catch (error) {
-      console.error(error);
-    } finally {
-      hide();
-    }
-  }
 
-  async function handleFlameGraph(app: AppListRecord) {
-    const hide = createMessage.loading('flameGraph generating...', 0);
-    try {
-      const { data } = await fetchFlamegraph({
-        appId: app.id,
-        width: document.documentElement.offsetWidth || document.body.offsetWidth,
-      });
-      if (data != null) {
-        const blob = new Blob([data], { type: 'image/svg+xml' });
-        const imageUrl = (window.URL || window.webkitURL).createObjectURL(blob);
-        window.open(imageUrl);
-      }
-    } catch (error) {
-      console.error(error);
-      createMessage.error('flameGraph generate failed');
-    } finally {
-      hide();
-    }
-  }
-
-  function handleCancel(app: AppListRecord) {
-    if (!optionApps.stopping.get(app.id) || app['optionState'] === 0) {
-      openStopModal(true, { application: app });
-    }
-  }
-  /* log view */
-  function handleSeeLog(app: AppListRecord) {
-    openLogModal(true, { app: unref(app) });
-  }
-  const {
-    handleCheckLaunchApp,
-    handleAppCheckStart,
-    handleCanStop,
-    handleForcedStop,
-    handleCopy,
-    handleMapping,
-    users,
-  } = useFlinkApplication(openStartModal);
-
-  /*  tag */
-  function handleInitTagsOptions() {
-    const params = Object.assign({}, { pageSize: 999999999, pageNum: 1 });
-    fetchAppRecord(params).then((res) => {
-      const dataSource = res?.records || [];
-      dataSource.forEach((record) => {
-        if (record.tags !== null && record.tags !== undefined && record.tags !== '') {
-          const tagsArray = record.tags.split(',') as string[];
-          tagsArray.forEach((x: string) => {
-            if (x.length > 0 && tagsOptions.value.indexOf(x) == -1) {
-              tagsOptions.value.push(x);
-            }
-          });
-        }
-      });
-    });
-  }
-
-  /* Operation button */
-  function getTableActions(record: AppListRecord): ActionItem[] {
-    return [
-      {
-        tooltip: { title: 'Edit Application' },
-        auth: 'app:update',
-        icon: 'clarity:note-edit-line',
-        onClick: handleEdit.bind(null, record),
-      },
-      {
-        tooltip: { title: 'Launch Application' },
-        ifShow: [-1, 1, 4].includes(record.launch) && record['optionState'] === 0,
-        icon: 'ant-design:cloud-upload-outlined',
-        onClick: handleCheckLaunchApp.bind(null, record),
-      },
-      {
-        tooltip: { title: 'Launching Progress Detail' },
-        ifShow: [-1, 2].includes(record.launch) || record['optionState'] === 1,
-        auth: 'app:update',
-        icon: 'ant-design:container-outlined',
-        onClick: openBuildProgressDetailDrawer.bind(null, record),
-      },
-      {
-        tooltip: { title: 'Start Application' },
-        ifShow: handleIsStart(record, optionApps),
-        auth: 'app:start',
-        icon: 'ant-design:play-circle-outlined',
-        onClick: handleAppCheckStart.bind(null, record),
-      },
-      {
-        tooltip: { title: 'Cancel Application' },
-        ifShow: record.state === 5 && record['optionState'] === 0,
-        auth: 'app:cancel',
-        icon: 'ant-design:pause-circle-outlined',
-        onClick: handleCancel.bind(null, record),
-      },
-      {
-        tooltip: { title: 'View Application Detail' },
-        auth: 'app:detail',
-        icon: 'ant-design:eye-outlined',
-        onClick: handleDetail.bind(null, record),
-      },
-      {
-        tooltip: { title: 'See Flink Start log' },
-        ifShow: [5, 6].includes(record.executionMode),
-        auth: 'app:detail',
-        icon: 'ant-design:sync-outlined',
-        onClick: handleSeeLog.bind(null, record),
-      },
-      {
-        tooltip: { title: 'Forced Stop Application' },
-        ifShow: handleCanStop(record),
-        auth: 'app:cancel',
-        icon: 'ant-design:pause-circle-outlined',
-        onClick: handleForcedStop.bind(null, record),
-      },
-    ];
-  }
-
-  /* pull down button */
-  function getActionDropdown(record: AppListRecord): ActionItem[] {
-    return [
-      {
-        label: 'Copy Application',
-        auth: 'app:copy',
-        icon: 'ant-design:copy-outlined',
-        onClick: handleCopy.bind(null, record),
-      },
-      {
-        label: 'Remapping Application',
-        ifShow: [0, 7, 10, 11, 13].includes(record.state),
-        auth: 'app:mapping',
-        icon: 'ant-design:deployment-unit-outlined',
-        onClick: handleMapping.bind(null, record),
-      },
-      {
-        label: 'View FlameGraph',
-        ifShow: record.flameGraph,
-        auth: 'app:flameGraph',
-        icon: 'ant-design:fire-outlined',
-        onClick: handleFlameGraph.bind(null, record),
-      },
-      {
-        popConfirm: {
-          title: 'Are you sure delete this job ?',
-          confirm: handleDelete.bind(null, record),
-        },
-        label: 'Delete',
-        ifShow: [0, 7, 9, 10, 13, 18, 19].includes(record.state),
-        auth: 'app:delete',
-        icon: 'ant-design:delete-outlined',
-        color: 'error',
-      },
-    ];
-  }
   /* Update options data */
   function handleOptionApp(data: {
     type: 'starting' | 'stopping' | 'launch';
@@ -398,7 +170,7 @@
 
   const { start, stop } = useTimeoutFn(() => {
     if (!getLoading()) {
-      handleDashboard(false);
+      appDashboardRef.value?.handleDashboard(false);
       reload({ polling: true });
     }
     start();
@@ -408,10 +180,6 @@
     stop();
   });
 
-  onMounted(() => {
-    handleDashboard(true);
-    handleInitTagsOptions();
-  });
   watch(
     () => [tags.value, userId.value, jobType.value],
     () => {
@@ -421,80 +189,8 @@
 </script>
 <template>
   <PageWrapper contentFullHeight>
-    <Row :gutter="24" class="dashboard">
-      <Col
-        class="gutter-row mt-10px"
-        :md="6"
-        :xs="24"
-        v-for="(value, key) in dashBigScreenMap"
-        :key="key"
-      >
-        <StatisticCard
-          :statisticProps="value.staticstics"
-          :footerList="value.footer"
-          :loading="dashboardLoading"
-        />
-      </Col>
-    </Row>
-    <BasicTable @register="registerTable" class="app_list !px-0 mt-20px">
-      <template #headerTop>
-        <div class="text-right my-15px">
-          <InputGroup compact>
-            <div class="pr-16px">
-              <Select
-                placeholder="Tags"
-                show-search
-                allowClear
-                v-model:value="tags"
-                class="!w-150px text-left"
-              >
-                <SelectOption v-for="tag in tagsOptions" :key="tag">{{ tag }}</SelectOption>
-              </Select>
-            </div>
-
-            <div class="pr-16px">
-              <Select
-                placeholder="Owner"
-                allowClear
-                v-model:value="userId"
-                class="!w-120px text-left"
-              >
-                <SelectOption v-for="u in users" :key="u.userId">
-                  <span v-if="u.nickName"> {{ u.nickName }} </span>
-                  <span v-else> {{ u.username }} </span>
-                </SelectOption>
-              </Select>
-            </div>
-            <div class="pr-16px">
-              <Select
-                placeholder="Type"
-                allowClear
-                v-model:value="jobType"
-                class="w-100px text-left"
-              >
-                <SelectOption value="1">JAR</SelectOption>
-                <SelectOption value="2">SQL</SelectOption>
-              </Select>
-            </div>
-            <div class="pr-16px">
-              <InputSearch
-                placeholder="Search..."
-                v-model:value="searchText"
-                @search="reload({ polling: true })"
-                class="!w-250px text-left"
-              />
-            </div>
-            <a-button
-              type="primary"
-              style="margin-left: 20px"
-              @click="router.push({ path: '/flink/app/add' })"
-            >
-              <Icon icon="ant-design:plus-outlined" />
-              {{ t('common.add') }}
-            </a-button>
-          </InputGroup>
-        </div>
-      </template>
+    <AppDashboard ref="appDashboardRef" />
+    <BasicTable @register="registerTable" class="app_list !px-0 mt-20px" :formConfig="formConfig">
       <template #bodyCell="{ column, record }">
         <template v-if="column.dataIndex === 'jobName'">
           <span class="app_type app_jar" v-if="record['jobType'] === 1"> JAR </span>
