+++ streampark-console/streampark-console-newui/src/views/flink/app/styles/Add.less
@@ -296,3 +296,9 @@
   transform: translateX(10px);
   opacity: 0;
 }
+
+.pom-card {
+  .ant-tabs-nav {
+    margin-bottom: 0 !important;
+  }
+}
