+++ streampark-console/streampark-console-newui/src/views/flink/notebook/Submit.less
@@ -14,6 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+[data-theme="dark"] {
+  .nodebook-submit {
+    .current-line,
+    .current-line-both {
+      border: 1px solid hsl(205deg 100% 74% / 20%) !important;
+      background-color: hsl(205deg 100% 74% / 20%);
+    }
+  }
+}
+
 .nodebook-submit {
   .code-box {
     height: 600px;
